Split up media3 module
