+++ media/media3-audiooffload/README.md
@@ -1,0 +1,17 @@
+# Wear Media3 Audio Offload library
+
+[![Maven Central](https://img.shields.io/maven-central/v/com.google.android.horologist/horologist-media3-audiooffload)](https://search.maven.org/search?q=g:com.google.android.horologist)
+
+For more information, visit the documentation: https://google.github.io/horologist/media3-audiooffload
+
+## Download
+
+```groovy
+repositories {
+    mavenCentral()
+}
+
+dependencies {
+    implementation "com.google.android.horologist:horologist-media3-audiooffload:<version>"
+}
+```
