+++ media/media3-audiooffload/gradle.properties
@@ -1,0 +1,3 @@
+POM_ARTIFACT_ID=horologist-media3-audiooffload
+POM_NAME=Horologist Media3 Audio Offload library
+POM_PACKAGING=aar
