+++ media/media3-logging/gradle.properties
@@ -1,0 +1,3 @@
+POM_ARTIFACT_ID=horologist-media3-logging
+POM_NAME=Horologist Media3 Logging library
+POM_PACKAGING=aar
