+++ media/media3-outputswitcher/gradle.properties
@@ -1,0 +1,3 @@
+POM_ARTIFACT_ID=horologist-media3-outputswitcher
+POM_NAME=Horologist Media3 Output Switcher library
+POM_PACKAGING=aar
