+++ settings.gradle.kts
@@ -47,6 +47,9 @@ include(":media:audio")
 include(":media:audio-ui")
 include(":media:core")
 include(":media:backend-media3")
+include(":media:media3-audiooffload")
+include(":media:media3-logging")
+include(":media:media3-outputswitcher")
 include(":media:ui")
 include(":media:benchmark")
 include(":media:data")
